requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-core 
langchain-community 
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv